Combined Teacher and Student into ApplicationUser.cs entity, update naming for lambda variable

+++ AnalysisofKnowledge.Database/Entities/Interfaces/IStudent.cs
@@ -2,6 +2,6 @@ namespace AnalysisofKnowledge.Database.Entities.Interfaces
 {
     public interface IStudent : IIdentityEntity
     {
-        int GroupId { get; set; }
+        long GroupId { get; set; }
     }
 }
